Suppressed warnings about deprecation where the replacement is not available
Because the target SDK version is Jelly Bean but the minimum SDK version is
Donut, all APIs deprecated in JB throw warnings, but some of them have no
alternatives that are available in Donut.

The Notification API used in NotificationHandler is replaced by
Notification.Builder in Honeycomb, and the Preference API used in
PreferencesActivity is replaced by PreferenceFragment, also in Honeycomb.