Renamed emergencyPageIntent to pageIntent
Part way through the pager functionality development, I decided against
labeling it as an emergency pager, just in case it failed to activate in
response to a message, the message never got to the device, or something else.
The pager functionality is a "best effort" mechanism, and should be treated as
such.